[CI/CD & util] Develop SDS, SkipList, LinkedList modules and modify the CI/CD configuration
